+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 ## [0.12.0] - 2022-04-09
 ### Added
 - `Wit` type [PR #25](https://github.com/bnjjj/witgen/pull/25)
-- `Witgen` struct and expose `cargo-witgen` as a library [#PR 26](https://github.com/bnjjj/witgen/pull/25)
+- `Witgen` struct; expose `cargo-witgen` as a library; and add new arg `input_dir` [#PR 26](https://github.com/bnjjj/witgen/pull/25)
 ### Changed
  - Use `syn-file-expand` instead of Cargo [PR #25](https://github.com/bnjjj/witgen/pull/25)
 ---